+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.3.13] - (2024-Feb-27)
+
+- Add `module` option in createSplash
+- Add `module` option in removeSplash
+
 ## [2.3.13] - (2024-Feb-18)
 
 - Added better exception handling based on [this comment](https://github.com/jonbhanson/flutter_native_splash/issues/643#issuecomment-1938416308).  Thanks [BujusKrachus](https://github.com/BujusKrachus) for the detailed explanation.